[common] A utility to extract MAC address from host (#894)
Summary: A utility to extract the MAC address from /sys/class/net. It
will ignore any virtual devices (like docker and br devices), but
otherwise use the lexicographically first device on the host.

Relevant Issues: GML-231

Type of change: /kind feature

Test Plan: A basic test runs the function. The test checks for physical
devices, so there is some risk that the test will behave differently on
different hosts. For example, if a host has no physical devices, this
test might fail. We can deal with that scenario if it ever comes to
light.

GitOrigin-RevId: 9248859259f48d9ce39566d37e34fe204ef0168b

+++ src/common/base/utils.h
@@ -225,7 +225,7 @@ inline StatusOr<T> AsciiHexToBytes(std::string s, const std::vector<char>& separ
 
   T bytes;
 
-  for (unsigned int i = 0; i < s.length(); i += 2) {
+  for (size_t i = 0; i < s.length(); i += 2) {
     std::string byte_string = s.substr(i, 2);
 
     errno = 0;
